refactoring the Google API spreadsheet
